Conflicts:
	CMakeLists.txt
	src/ctest_python.py
	src/frontend/synergia.in
	src/synergia/__init__.py
	src/synergia/bunch/CMakeLists.txt
	src/synergia/bunch/bunch_wrap.cc
	src/synergia/bunch/diagnostics.h
	src/synergia/bunch/period.h
	src/synergia/bunch/tests/CMakeLists.txt
	src/synergia/collective/CMakeLists.txt
	src/synergia/foundation/CMakeLists.txt
	src/synergia/lattice/CMakeLists.txt
	src/synergia/simulation/CMakeLists.txt
	src/synergia/simulation/lattice_simulator.cc
	src/synergia/synergia.cc.in
	src/synergia/tools/three_bump.py
	src/synergia/utils/CMakeLists.txt
	src/synergia/version.py.in
	synergia-local.in

+++ src/synergia/bunch/bunch.cc
@@ -804,7 +804,7 @@ void do_sort(double * t, size_t rows, size_t cols, size_t cols_padded, size_t or
     // c++ 11
     // unsigned int ind=0;
     //generate(index.begin(),index.end(), [&]() { return ind++; });
-    for(int i=0; i<cols; ++i) {
+    for(size_t i=0; i<cols; ++i) {
         index[i] = i;
     }
     semi_global_start_pos = ord_col*cols_padded;
